py3 friendly dict handling
dict keys() and values() retruns with generator in python3,
if you would like to have the first element only (you know it has just
1 element) than converting it to the list is the fastest way.
In case of long lists the next(iter(D.keys())) might be an option.

iteritems was generally good idea to use for iterating over big dicts in
a loop, but iteritems is not dict member in py3, but items() behaves in a
same way in py3. Using items() in py2 gives you list, which might not
be a good idea when your dict consumes multiple megabytes.
Anyway the usual recommended way for being py3 ready is using items().

Change-Id: I52ae8a92369d168825cc047495516be957d3abbb